@@ -18,17 +18,16 @@
         }
     ],
     "require": {
-        "php":                  "^7.3.0",
-        "doctrine/orm":         "^2.7.4",
+        "php":                  "^7.4.0|| ^8.0",
+        "doctrine/orm":         "^2.8",
         "doctrine/persistence": "^1.3.3 || ^2.0"
     },
     "require-dev": {
         "doctrine/coding-standard": "^8.2.0",
-        "infection/infection": "^0.15.2",
-        "phpstan/phpstan": "^0.12.52",
-        "phpstan/phpstan-phpunit": "^0.12.16",
-        "phpunit/phpunit": "^9.2.6",
-        "roave/backward-compatibility-check": "^4.4",
+        "infection/infection": "^0.20.2",
+        "phpstan/phpstan": "^0.12.61",
+        "phpstan/phpstan-phpunit": "^0.12.14",
+        "phpunit/phpunit": "^9.5",
         "vimeo/psalm": "^4.0.1"
     },
     "autoload": {
